pet-care-assistant/requirements.txt
fastapi
uvicorn
langchain
chromadb
python-dotenv
sentence-transformers
langchain_community
langchain_google_genai
